Refactor Requestor classes
Add BaseRequestor and refactor existing requestors to inherit from it
